+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `RehauNeaSmartIoDevice` class in `pyrehau_neasmart/iodevice.py`.
 - Added `iodevices()` method to `RehauNeaSmart` to list IO devices.
 - Added `get_heatarea(nr)` method to `RehauNeaSmart`.
+- Added `get_iodevice(nr)` method to `RehauNeaSmart`.
+